pyproject.toml
[tool.poetry]
name = "atlantisbot"
version = "0.1.0"
description = ""
authors = ["johnvictorfs <johnvictorfs@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.3.4"
rs3clans = "^1.1"
beautifulsoup4 = "^4.7"
colorama = "^0.4.1"
matplotlib = "^3.0"
lxml = "^4.3"
psycopg2-binary = "^2.8"
requests = {extras = ["security"], version = "^2.23.0"}
feedparser = "^5.2"
python-twitter = "^3.5"
Pillow = "^6.1"
aioify = "^0.3.2"
pandas = "^1.1.4"
dj-database-url = "^0.5.0"
atlantisbot-api = "^1.0.5"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
autopep8 = "^1.4"
mypy = "^0.740.0"
mypy_extensions = "^0.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
